src/requirements.txt
Django>=5.0
psycopg[binary]
djongo
pymongo
djangorestframework